Add ability to set up a puzzle from a text description
Hey Sadie, in order to test this out I added a bit of code which you might find useful if you ever want to work on this again. As you can see I made some small adjustments to `updateBoard` so I could use it in a new function `setupPuzzle` that does what it says on the tin. (I got tired of clicking in puzzles by hand.)

I also added a brute-force throttle to the while loop around `solvePuzzle` in case the solver isn't making progress. The point of all of this was to test whether your solver works for all puzzles because I wasn't sure it would. And, as you can see if you switch to the `hard` puzzle it can get stuck when the simple constraints you're checking aren't sufficient.

If you want to learn more about this https://norvig.com/sudoku.html is a great essay by Peter Norvig of Google about solving sudoku.

I'm going to request *your* review on this PR and leave comments/ask questions or whatever you want. If you choose "Approve" then I'll merge it. (This is the normal Github workflow most places where the author of the code is responsible for merging it after it has been approved.) Though if you really wanted to start building on this code you could also merge it yourself.